+++ tests/xmlReader_test.js
@@ -100,7 +100,7 @@ describe('XML Reader', () => {
       ])
     })
 
-    expect(jsonData.tests[0].title).to.eql("login with valid credentials")
+    expect(jsonData.tests[0].title).to.eql("Login With Valid Credentials")
 
     const failedTests = jsonData.tests.filter(t => t.status === 'failed')
   })
@@ -134,7 +134,7 @@ describe('XML Reader', () => {
 
     expect(jsonData.tests[0].code).to.be.ok;
     expect(jsonData.tests[0].code).to.include("public function runCestWithTwoFailedTest(")
-    expect(jsonData.tests[0].title).to.eql("run cest with two failed test")
+    expect(jsonData.tests[0].title).to.eql("Run Cest With Two Failed Test")
 
     const failedTests = jsonData.tests.filter(t => t.status === 'failed')
     const failedTest = failedTests[0];
@@ -142,7 +142,7 @@ describe('XML Reader', () => {
   })  
 
   it('should parse JUnit XML', () => {
-    const reader = new XmlReader();
+    const reader = new XmlReader({ lang: 'java' });
     const jsonData = reader.parse(path.join(__dirname, 'data/java.xml'))
 
     expect(jsonData.status).to.eql('failed')
@@ -174,6 +174,7 @@ describe('XML Reader', () => {
 
     const failedTests = jsonData.tests.filter(t => t.status === 'failed')
     const failedTest = failedTests[0];
+    console.log(failedTest.stack)
     expect(failedTest.stack).to.include('(CalculatorTest.java:43')
   })
 
@@ -205,8 +206,38 @@ describe('XML Reader', () => {
     const skippedTests = jsonData.tests.filter(t => t.status === 'skipped')
     expect(skippedTests.length).to.eql(1)
     const skippedTest = skippedTests[0];
-    expect(skippedTest.title).to.eql('check dashboard')
-  })  
+    expect(skippedTest.title).to.eql('Check Dashboard')
+  })
+
+  it('should parse JUnit params', () => {
+    const reader = new XmlReader({ lang: 'java' });
+    const jsonData = reader.parse(path.join(__dirname, 'data/junit2.xml'))
+
+    expect(jsonData.status).to.eql('passed')
+    const stats = reader.calculateStats();
+    expect(stats.status).to.eql('passed')
+    expect(stats.tests_count).to.eql(4)
+
+    reader.connectAdapter();
+    reader.fetchSourceCode();
+    reader.formatErrors();
+    reader.formatTests();
+
+    jsonData.tests.forEach(t => {
+      expect(t.title).to.eql('Can Create API Key ${param}')
+      expect(t).to.contain.keys([
+        'stack',
+        'create',
+        'status',
+        'file',
+        'title',
+        'run_time',
+        'suite_title',
+      ])
+    })
+
+    expect(jsonData.tests[0].example.param).to.eql('Master');
+  })    
 
 
   it('should parse JUnit C#', () => {
@@ -237,7 +268,7 @@ describe('XML Reader', () => {
     })
     
     const tests = jsonData.tests;
-    expect(tests[0].title).to.include('Create a web lead');
+    expect(tests[0].title).to.include('Create a Web Lead');
     expect(tests[0].suite_title).to.include('User');
   })
 
@@ -263,7 +294,7 @@ describe('XML Reader', () => {
     })
     
     const tests = jsonData.tests;
-    expect(tests[0].title).to.include('Create user login');
+    expect(tests[0].title).to.include('Create User Login');
     expect(tests[0].suite_title).to.include('User');
   })
 
@@ -289,7 +320,7 @@ describe('XML Reader', () => {
     })
     
     const tests = jsonData.tests;
-    expect(tests[0].title).to.include('Update user');
+    expect(tests[0].title).to.include('Update User');
     expect(tests[0].suite_title).to.include('User');
   })  
 
@@ -315,7 +346,7 @@ describe('XML Reader', () => {
     })
     
     const tests = jsonData.tests;
-    expect(tests[0].title).to.include('Test method1');
+    expect(tests[0].title).to.include('Method1');
     expect(tests[0].suite_title).to.include('TestClass1');
     expect(tests[0].file).to.eql('Sample/Tests');
     expect(tests[0].status).to.eql('passed');